Refactor/split panels into folders
Pure file reorganization. Behavior unchanged.

- Each panel domain gets its own folder under `panels/`: `edge/`,
  `node/`, `room/`.
- `node-metadata-panel.tsx` and `room-metadata-panel.tsx` are split
  into separate Create + Edit + shared-helpers + dispatcher files.
- Dispatchers renamed to `*-panels.tsx` (`NodePanels`, `RoomPanels`,
  `EdgePanel`).
- Route imports updated to the new paths.
