+++ src/components/panels/room/room-info-panel.tsx
